Añadido test unitarios

+++ src/main/java/com/vector/tests/wallet/entities/Monedero.java
@@ -1,5 +1,7 @@
 package com.vector.tests.wallet.entities;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -17,7 +19,7 @@ public class Monedero {
     private String identificador;
 
     @Column(name = "saldo", nullable = false)
-    private Integer saldo;
+    private BigDecimal saldo;
 
 	public String getIdentificador() {
 		return identificador;
@@ -35,11 +37,11 @@ public class Monedero {
 		this.id = id;
 	}
 
-	public Integer getSaldo() {
+	public BigDecimal getSaldo() {
 		return saldo;
 	}
 
-	public void setSaldo(Integer saldo) {
+	public void setSaldo(BigDecimal saldo) {
 		this.saldo = saldo;
 	}
 }
